Rework filter defaults
This fixes the apischema generation for the HyperlinkRelatedFilter.

fixes #3351
