+++ CHANGES/3351.bugfix
@@ -1,0 +1,1 @@
+Fixes schema generation to use proper type `uri` for ``HyperlinkRelatedFilter``.
