+++ pulpcore/app/viewsets/access_policy.py
@@ -1,28 +1,14 @@
-from django_filters.rest_framework import filters
 from rest_framework import mixins
 from rest_framework.decorators import action
 from rest_framework.response import Response
 from drf_spectacular.utils import extend_schema
 
-from pulpcore.filters import BaseFilterSet
 from pulpcore.app.apps import pulp_plugin_configs
 from pulpcore.app.models import AccessPolicy
 from pulpcore.app.serializers import AccessPolicySerializer
 from pulpcore.app.viewsets import NamedModelViewSet
 from pulpcore.app.viewsets.base import NAME_FILTER_OPTIONS
 from pulpcore.app.util import get_view_urlpattern
-
-
-class AccessPolicyFilter(BaseFilterSet):
-    """
-    FilterSet for AccessPolicy.
-    """
-
-    customized = filters.BooleanFilter()
-
-    class Meta:
-        model = AccessPolicy
-        fields = {"viewset_name": NAME_FILTER_OPTIONS, "customized": ["exact"]}
 
 
 class AccessPolicyViewSet(
@@ -39,7 +25,7 @@ class AccessPolicyViewSet(
     serializer_class = AccessPolicySerializer
     endpoint_name = "access_policies"
     ordering = "viewset_name"
-    filterset_class = AccessPolicyFilter
+    filterset_fields = {"viewset_name": NAME_FILTER_OPTIONS, "customized": ["exact"]}
 
     @extend_schema(request=None)
     @action(detail=True, methods=["post"])
